Consolidate AzureResourceGroup into Calamari.Azure
