+++ source/Calamari.Azure/Calamari.Azure.csproj
@@ -15,6 +15,7 @@
         <PackageReference Include="Azure.ResourceManager" Version="1.11.0"/>
         <PackageReference Include="Azure.ResourceManager.AppService" Version="1.0.2"/>
         <PackageReference Include="Azure.ResourceManager.ResourceGraph" Version="1.0.1"/>
+        <PackageReference Include="Azure.ResourceManager.Resources" Version="1.7.0" />
         <PackageReference Include="Microsoft.AspNetCore.WebUtilities" Version="2.2.0"/>
         <PackageReference Include="Microsoft.DotNet.Analyzers.Compatibility" Version="0.2.12-alpha">
             <PrivateAssets>all</PrivateAssets>
@@ -32,6 +33,7 @@
         </PackageReference>
     </ItemGroup>
     <ItemGroup>
+        <ProjectReference Include="..\Calamari.AzureScripting\Calamari.AzureScripting.csproj" />
         <ProjectReference Include="..\Calamari.CloudAccounts\Calamari.CloudAccounts.csproj"/>
         <ProjectReference Include="..\Calamari.Common\Calamari.Common.csproj"/>
     </ItemGroup>
