+++ source/Calamari.Azure/InternalsVisibleTo.cs
@@ -1,3 +1,4 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Calamari.AzureAppService.Tests")]
+[assembly: InternalsVisibleTo("Calamari.AzureResourceGroup.Tests")]
