+++ source/Calamari.Azure/Program.cs
@@ -1,6 +1,13 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.Reflection;
+using System.Threading.Tasks;
+using Autofac;
+using Calamari.Azure.ResourceGroups;
+using Calamari.AzureScripting;
 using Calamari.Common;
+using Calamari.Common.Plumbing.Commands;
 using Calamari.Common.Plumbing.Logging;
+using Calamari.Common.Util;
 
 namespace Calamari.Azure
 {
@@ -9,6 +16,23 @@ namespace Calamari.Azure
         public Program(ILog log)
             : base(log)
         { }
+
+        protected override void ConfigureContainer(ContainerBuilder builder, CommonOptions options)
+        {
+            base.ConfigureContainer(builder, options);
+
+            builder.RegisterType<TemplateService>();
+            builder.RegisterType<ResourceGroupTemplateNormalizer>().As<IResourceGroupTemplateNormalizer>();
+            builder.RegisterType<TemplateResolver>().As<ITemplateResolver>().SingleInstance();
+            builder.RegisterType<AzureResourceGroupOperator>();
+            
+        }
+        
+        protected override IEnumerable<Assembly> GetProgramAssembliesToRegister()
+        {
+            yield return typeof(AzureContextScriptWrapper).Assembly;
+            yield return typeof(Program).Assembly;
+        }
 
         public static Task<int> Main(string[] args)
         {
