--- source/Calamari.AzureResourceGroup/Properties/InternalsVisibleTo.cs
@@ -1,3 +1,0 @@
-using System.Runtime.CompilerServices;
-
-[assembly:InternalsVisibleTo("Calamari.AzureResourceGroup.Tests")]
