+++ source/Calamari.sln
@@ -48,8 +48,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.ConsolidateCalamar
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.ConsolidateCalamariPackages", "Calamari.ConsolidateCalamariPackages\Calamari.ConsolidateCalamariPackages.csproj", "{5A02308A-139B-417B-8FF7-1E14A3D175B5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.AzureResourceGroup", "Calamari.AzureResourceGroup\Calamari.AzureResourceGroup.csproj", "{8FCEDC44-EC0C-4325-A342-149C469F5CA2}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.AzureResourceGroup.Tests", "Calamari.AzureResourceGroup.Tests\Calamari.AzureResourceGroup.Tests.csproj", "{0EA7711F-34D3-44C6-BBCC-D4EA1A6B9D2C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.AzureWebApp", "Calamari.AzureWebApp\Calamari.AzureWebApp.csproj", "{BFE038E8-6F46-40DA-AD40-C00ED25C1FEE}"
@@ -134,10 +132,6 @@ Global
 		{5A02308A-139B-417B-8FF7-1E14A3D175B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5A02308A-139B-417B-8FF7-1E14A3D175B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5A02308A-139B-417B-8FF7-1E14A3D175B5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8FCEDC44-EC0C-4325-A342-149C469F5CA2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8FCEDC44-EC0C-4325-A342-149C469F5CA2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8FCEDC44-EC0C-4325-A342-149C469F5CA2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8FCEDC44-EC0C-4325-A342-149C469F5CA2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0EA7711F-34D3-44C6-BBCC-D4EA1A6B9D2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0EA7711F-34D3-44C6-BBCC-D4EA1A6B9D2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0EA7711F-34D3-44C6-BBCC-D4EA1A6B9D2C}.Release|Any CPU.ActiveCfg = Release|Any CPU
